pyproject.toml
[tool.poetry]
name = "notes-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.5.8"
appdirs = "^1.4.4"
attrs = "^20.3.0"
black = "^21.4b2"
certifi = "^2020.12.5"
cfgv = "^3.2.0"
chardet = "^4.0.0"
click = "^7.1.2"
distlib = "^0.3.1"
fastapi = "^0.87.0"
filelock = "^3.0.12"
h11 = "^0.12.0"
identify = "^2.2.4"
idna = "^2.10"
inflect = "^5.3.0"
iniconfig = "^1.1.1"
mako = "^1.1.4"
markupsafe = "^1.1.1"
mypy-extensions = "^0.4.3"
nodeenv = "^1.6.0"
packaging = "^20.9"
pathspec = "^0.12.1"
pluggy = "^0.13.1"
pre-commit = "^2.12.1"
py = "^1.10.0"
pydantic = {extras = ["email"], version = "^1.10.2"}
pyparsing = "^2.4.7"
pytest = "^7.1.3"
python-dateutil = "^2.8.1"
python-dotenv = "^0.17.1"
python-editor = "^1.0.4"
pyyaml = "^6.0.2"
regex = "^2021.4.4"
requests = "^2.25.1"
six = "^1.15.0"
sqlalchemy = "^2.0.34"
sqlalchemy-utils = "^0.41.2"
starlette = "^0.21.0"
toml = "^0.10.2"
typing-extensions = "^4.1.0"
urllib3 = "^1.26.4"
uvicorn = "^0.13.4"
virtualenv = "^20.4.4"
passlib = "^1.7.4"
factory-boy = "^3.2.1"
pytest-env = "^0.8.1"
pyjwt = "^2.6.0"
python-decouple = "^3.6"
fastapi-pagination = "^0.10.0"
httpx = "^0.23.1"
bcrypt = "^4.0.1"
psycopg2-binary = "^2.9.5"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
ruff = "^0.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
env = [
    "TESTING=True",
]

[tool.black]
line-length = 100

[tool.ruff]
exclude = [".venv", "alembic"]
